test(test_download.py): check old/new keyword and stereochemistry
The edit checks both usage of the depreciated keyword of canonical
SMILES (intentionally retained for backward compatibility) as well
as for the contemporary keyword of connectivity SMILES.  The tests
now equally account for stereochemistry in the more general output
of `SMILES` output which can yield a different string to one about
connectivity alone -- for instance if the compound contains a
stereogenic center.  To probe for the later, the example added is
again CID5950, (S)-alanine.

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>